+++ projects/rocprofiler-compute/src/lib/CMakeLists.txt
@@ -11,5 +11,8 @@ endif()
 find_package(rocprofiler-sdk REQUIRED)
 
 add_library(rocprofiler-compute-tool SHARED)
-target_sources(rocprofiler-compute-tool PRIVATE rocprofiler_compute_tool.cpp)
+target_sources(rocprofiler-compute-tool PRIVATE 
+    rocprofiler_compute_tool.cpp
+    helper.cpp
+)
 target_link_libraries(rocprofiler-compute-tool PRIVATE rocprofiler-sdk::rocprofiler-sdk)
